requirements.txt
streamlit>=1.28.0
watchdog>=3.0.0
librosa>=0.10.0
ffmpeg-python>=0.2.0
pandas>=2.0.0
google-generativeai>=0.3.0
numpy>=1.24.0
Pillow>=10.0.0

